rename patch path_changes to make its name more meaningful
we have way too many patches with names that imply they do the same
thing. This patch is about javascript files so let's add this info in
the name

bonus: added metadata to show that the patch does not need to be
forwarded upstream since this is just a difference in how things are
done in debian vs upstream.