merge both patches that hardcode config path
Both of these patches do the same thing, but one is only working on the
cgi executable.

Both of the patches are not meant to be forwarded upstream.

One patch doesn't have headers.

Let's merge them: it'll be more logical this way and we'll have one less
patch to re-evaluate when packaging new upstream releases.