init.d: match on pidfile only prevents service stop (closes: #922395)
dpkg 1.19.3 introduced a new verification that prevents the service to
stop using the init script (and incidentally to run reconfigure
correctly): the pidfile is not owned by root:root, so s-s-d is wary of
relying only on this information to run a command as root with the
information within the file.

By matching the service more precisely, that is with pidfile and exec
name, start-stop-daemon is more at ease and performs service stop
action.

While changing that command, I also saw that the init script was not
reusing the $PIDFILE variable for the stop command. This could have
been a source of mistake if we ever had to change where the pidfile is
placed on the filesystem.