Use our alias to find css and js files (closes: #929515)
Some folks are still using URLs starting with /cgi-bin/ instead of the
new /smokeping alias. This is quite understandable since the alias is
relatively new.

Currently the css and js files are not loaded when using URLs starting
with /cgi-bin/ since the paths in basepage.html are relative. If we
reuse our top-level /smokeping alias we can make those files load in
all cases.