remove myself from uploaders
I do not intend to work on this package on my own for the forseeable
future. Last time I setup something like smokeping, I used prometheus
and grafana, and it worked fine, even if not as fine as this, it was
good enough.

I'm happy to sponsor uploads in the future, that said, but for that I
don't need to be maintainer.

Good luck gab! :)