Revert "fix lintian warning about unmatched news entry"
This reverts commit 668bf158793d23350efb1f99c60405d5e670d446.

That approach is not good and it doesn't even fix the lintian warning.

I was told by Myon on #debian-mentors that the lintian warning about
the version tag is a known debhelper bug where it trims version tags in
changelog but not in NEWS. I'll just ignore this and wait for debhelper
to get fixed.