lintian-overrides are now obsolete
Those overrides are not needed anymore since the repack removes those
vendored files from the upstream archive.

Note: the date of the commit might look off. That's because this change
used to live at this point in the history mixed in with the actual
repack. However, the repack was done too late (after the new upstream
release was already imported) and we're trying to fix this situation
while retaining all changes in place.