Update README wrt fcgi
With the recent merger of Sebastian Badia's suggested change for using
fcgid by default, the readme doesn't need the extra mention anymore.

We currently already have an example config file for lighttpd so we
should mention its existence in the README file. Another example config
file is also being worked on for nginx (in MR 5 on salsa)